broke up into components
