+++ README.md
@@ -7,3 +7,8 @@ npm install @amcharts/amcharts4
 npm install @amcharts/amcharts4-geodata
 
 before it'll work
+
+Plus, remember it's already been deployed to heroku.com, so to start it you don't use npm start any more, you use 
+
+##npm run dev##
+
